fix: full audit — pre-auth 50MB parse, register race, bodiless PUT 500s, coercion, mobile nav, blank PNGs
Server
- Global JSON parser skips /api/backup/restore so the 50MB parser only runs
  behind backup.js auth; anonymous clients no longer get large bodies parsed.
- Registration claims the single slot synchronously after hashing; two
  concurrent registers admit exactly one account.
- Register no longer writes ALLOW_REGISTRATION into the source-tree .env from
  tests; it writes the cwd .env dotenv actually reads, and skips under test.
- req.body defaults to {} (Express 5 leaves it undefined), so a bodiless PUT
  is a 400 instead of a TypeError 500 on seven handlers.
- Legacy databases with duplicate (class_id, date, start_time) rows are
  deduped before idx_schedules_unique is created instead of crashing initDb.
- Validators coerce: weekday/classIds/tier bounds .toInt(), boolean flags
  .toBoolean() (dryRun:"false" really deletes; isCompetition:"false" is
  false; classIds ["3"] no longer wipes a student's class links); string
  fields .isString().trim() so objects fail at 400 and whitespace names are
  rejected.
- POST /schedules inherits the class default location whole or not at all;
  naming another place no longer borrows the default's coordinates. Batch
  update clearing locationName clears its coordinates like the single update.
- Wrong current password is a 400: 401 made the client drop the session.
- Screenshot scale factor is lowered until the capture fits Chromium's
  raster budget; a 4x monthly export went blank after ~7 months. page.close
  failures no longer mask the real error.

Frontend
- navTarget takes the visible span (2 days on mobile) and, without a stored
  period, picks the month/year holding most of the week.
- Keyboard shortcuts stand down inside open dialogs and with Alt held.
- Editing a schedule shows auto billing as empty so time edits recompute it.
- Pricing edit can clear the discount; holiday add no longer prefills an
  invalid date; July defaults to the summer semester template.
- Month/year navigation replaces history entries; schedule cache clears on
  login; swipe hook ignores stale multi-touch state; simple swipe requires a
  horizontal single-finger gesture; blocks starting above the grid are
  clipped instead of drawn at full height; report currency shows 2 decimals.

Tests: 45 new unit tests (724 total), Playwright 126/126.

+++ server/validations/pricing-tiers.js
@@ -1,8 +1,8 @@
 import { body } from 'express-validator';
 
 export const validateCreateTier = [
-  body('minStudents').isInt({ min: 1 }).withMessage('最小人数须为正整数'),
-  body('maxStudents').isInt({ min: 1 }).withMessage('最大人数须为正整数')
+  body('minStudents').isInt({ min: 1 }).toInt().withMessage('最小人数须为正整数'),
+  body('maxStudents').isInt({ min: 1 }).toInt().withMessage('最大人数须为正整数')
     .custom((max, { req }) => {
       if (req.body.minStudents != null && +max < +req.body.minStudents) {
         throw new Error('最大人数须不小于最小人数');
@@ -13,8 +13,8 @@ export const validateCreateTier = [
 ];
 
 export const validateUpdateTier = [
-  body('minStudents').optional().isInt({ min: 1 }).withMessage('最小人数须为正整数'),
-  body('maxStudents').optional().isInt({ min: 1 }).withMessage('最大人数须为正整数')
+  body('minStudents').optional().isInt({ min: 1 }).toInt().withMessage('最小人数须为正整数'),
+  body('maxStudents').optional().isInt({ min: 1 }).toInt().withMessage('最大人数须为正整数')
     .custom((max, { req }) => {
       if (req.body.minStudents != null && +max < +req.body.minStudents) {
         throw new Error('最大人数须不小于最小人数');
